commit before merging
