--- client/src/lib/artistdata.js
@@ -1,35 +1,0 @@
-const artistdata = [
-    {
-      artist_name: 1,
-      artistname: 'The Roots',
-      infos:
-        'Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam.',
-      tracks: [51]
-    },
-   
-    {
-      artist_name: 2,
-      artistname: 'The Pharcyde',
-      infos:
-        'Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam.',
-      tracks: [52]
-    },
-    {
-      artist_name: 4,
-      artistname: 'Drake',
-      infos:
-        'Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam.',
-      tracks: [53]
-    },
-
-    {
-      artist_name: 5,
-      artistname: 'Quelle Chris',
-      infos:
-        'Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam.',
-      tracks: [54]
-    },
-
-  ]
-  
-  export default artistdata
