--- client/src/lib/tracksdata.js
@@ -1,144 +1,0 @@
-export default [
-    {
-      track_id: '51',
-      artist_name: 'The Roots',
-      title: 'You Got Me',
-      album: 'Things Fall Apart',
-      year: 1998,
-      sampled_in: ['53'],
-      sampled: ['52'],
-      video_id: 'MJCHeEQV454',
-      cover_image: 'https://img.discogs.com/6KausygjQhBG8kIDUMurgbZ4qy4=/fit-in/600x597/filters:strip_icc():format(jpeg):mode_rgb():quality(90)/discogs-images/R-1000276-1183517792.jpeg.jpg'
-    },
-  
-    {
-      track_id: '52',
-      artist_name: 'The Pharcyde',
-      title: 'Passin Me By',
-      album: 'Bizarre Ride II the Pharcyde',
-      year: 1992,
-      sampled_in: ['51', '54', '57'],
-      sampled: [],
-      video_id: 'a-mAK3uB2_0',
-      cover_image: 'https://img.discogs.com/GAYgt__7N5BiU-JsHEx8tr8SmA4=/fit-in/414x416/filters:strip_icc():format(jpeg):mode_rgb():quality(90)/discogs-images/R-110360-1237804595.jpeg.jpg'
-    },
-  
-    {
-      track_id: '53',
-      artist_name: 'Drake',
-      title: '6 Man',
-      album: 'If You\'re Reading This It\'s Too Late',
-      year: 2015,
-      sampled_in: [],
-      sampled: ['51', '55'],
-      video_id: 'WvXTpZawgSM',
-      cover_image: 'https://img.discogs.com/5_eznyfDuZfReQj1e5szGXqR-xg=/fit-in/300x300/filters:strip_icc():format(jpeg):mode_rgb():quality(40)/discogs-images/R-7720513-1447414388-9059.jpeg.jpg'
-    },
-    {
-      track_id: '54',
-      artist_name: 'Drake',
-      title: 'Too Good',
-      album: 'If You\'re Reading This It\'s Too Late',
-      year: 2015,
-      sampled_in: [],
-      sampled: ['51', '55'],
-      video_id: '1lbKUiZv1fI',
-      cover_image: 'https://img.discogs.com/oAAInexBUvma6zosDOQa1rKSnh0=/fit-in/600x600/filters:strip_icc():format(jpeg):mode_rgb():quality(90)/discogs-images/R-10514711-1584388844-2825.jpeg.jpg'
-    },
-    {
-      track_id: '55',
-      artist_name: 'Public Enemy',
-      title: 'Shut \'Em Down',
-      album: 'If You\'re Reading This It\'s Too Late',
-      year: 1991,
-      sampled_in: ['53', '54'],
-      sampled: [],
-      video_id: 'e3p1SB_j-2E',
-      cover_image: 'https://img.discogs.com/jVrjwoeEopR2qztKMzvY1ggu0aI=/fit-in/600x600/filters:strip_icc():format(jpeg):mode_rgb():quality(90)/discogs-images/R-1537765-1328812613.jpeg.jpg'
-    },
-  
-    {
-      track_id: '56',
-      artist_name: 'Quincy Jones',
-      title: 'Summer in the City',
-      album: 'If You\'re Reading This It\'s Too Late',
-      year: 1973,
-      sampled_in: ['52'],
-      sampled: [],
-      video_id: 'jpbK4tLQH_s',
-      cover_image: 'https://www.whosampled.com/static/track_images_200/lr2_20081112_115939840289.jpg'
-    },
-    {
-      track_id: '57',
-      artist_name: 'Quelle Chris',
-      title: 'Passin\' Q By',
-      album: 'If You\'re Reading This It\'s Too Late',
-      year: 2016,
-      sampled_in: [],
-      sampled: ['52'],
-      video_id: 'a-mAK3uB2_0',
-      cover_image: 'https://www.whosampled.com/static/track_images_200/lr156822_2018119_16594486990.jpg'
-    },
-  
-    // Erykah Badu sampled by: 59 / contains samples of:
-    {
-      track_id: '58',
-      artist_name: 'Erykah Badu',
-      title: 'On & On',
-      album: 'Baduizm',
-      year: 1997,
-      sampled_in: ['59'],
-      sampled: [''],
-      video_id: '-CPCs7vVz6s',
-      cover_image: 'https://img.discogs.com/8NFNNJF3FiBkTs545-l8mqq4qQM=/fit-in/600x594/filters:strip_icc():format(jpeg):mode_rgb():quality(90)/discogs-images/R-905999-1231881488.jpeg.jpg'
-    },
-  
-  // Mos Def sampled:  / contains samples of: 58,
-    {
-      track_id: '59',
-      artist_name: 'Mos Def',
-      title: 'Mathematics',
-      album: 'Black on Both Sides',
-      year: 1999,
-      sampled_in: [''],
-      sampled: ['58', '60'],
-      video_id: 'm5vw4ajnWGA',
-      cover_image: 'https://img.discogs.com/yxkPzZ2TAy_sS5xcVVWfGfhYCwI=/fit-in/600x593/filters:strip_icc():format(jpeg):mode_rgb():quality(90)/discogs-images/R-4948161-1380295643-7410.jpeg.jpg'
-    },
-  
-    // James Brown sampled by: 59, / contains samples of:
-    {
-      track_id: '60',
-      artist_name: 'James Brown',
-      title: 'Funky Drummer',
-      album: 'In the Jungle Groove',
-      year: 1970,
-      sampled_in: ['59'],
-      sampled: [''],
-      video_id: 'AoQ4AtsFWVM',
-      cover_image: 'https://i.scdn.co/image/ab67616d0000b2731c99d951689d6d80f4a4d903'
-    },
-  
-  // Burial 
-    {
-      track_id: '61',
-      artist_name: 'Burial',
-      title: 'Chemz',
-      album: 'In the Jungle Groove',
-      year: 2020,
-      sampled_in: [],
-      sampled: ['62', '63'],
-      video_id: 'vGEy1nBUBuE',
-      cover_image: 'https://img.discogs.com/OJbl29AfxHYD3erRSWF-My_eOgc=/fit-in/600x600/filters:strip_icc():format(jpeg):mode_rgb():quality(90)/discogs-images/R-16963743-1610840148-1543.jpeg.jpg'
-    },
-    {
-      track_id: '62',
-      artist_name: 'Allure',
-      title: 'Head over Heels',
-      album: 'Allure',
-      year: 1997,
-      sampled_in: ['61'],
-      sampled: [''],
-      video_id: 'KJA3crID7aY',
-      cover_image: 'https://img.discogs.com/4YUyTeuj5gbDcIJgDKkFn0sqFaw=/fit-in/600x599/filters:strip_icc():format(jpeg):mode_rgb():quality(90)/discogs-images/R-541254-1278332244.jpeg.jpg'
-    }]
