--- client/src/lib/userdata.js
@@ -1,18 +1,0 @@
-export default [{
-    name: 'Jane',
-    last_name: 'Doe',
-    user_name: 'Super-Jane',
-    image:
-      'https://images.unsplash.com/photo-1479936343636-73cdc5aae0c3?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1160&q=80',
-    track_entries: [],
-  },
-{
-  name: 'John',
-  last_name: 'Doe',
-  user_name: 'Super-John',
-  image:
-    'https://images.unsplash.com/photo-1598362471594-d1f1c97a155d?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1160&q=80',
-  track_entries: [],
-},
-
-]
